Code-for-Experiment/Lyric-Generation-LLMs/lyric_cleaning/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
distro==1.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
multidict==6.0.4
mysql-connector-python==8.2.0
numpy==1.25.2
openai==1.3.2
openpyxl==3.1.2
pandas==2.0.3
pandas-stubs==2.0.2.230605
protobuf==4.21.12
pydantic==2.5.1
pydantic_core==2.14.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tqdm==4.65.0
types-pytz==2023.3.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
zhconv==1.4.3
